+++ packages/sampler/stories/moonstone-stories/ExpandableList.js
@@ -2,29 +2,32 @@ import ExpandableList, {ExpandableListBase} from '@enact/moonstone/ExpandableLis
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, select, text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
+import {boolean, select, text} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
+
+const Config = mergeComponentMetadata('ExpandableList', ExpandableList, ExpandableListBase);
+ExpandableList.displayName = 'ExpandableList';
 
 storiesOf('Moonstone', module)
 	.add(
 		'ExpandableList',
 		withInfo({
-			propTables: [ExpandableListBase],
+			propTablesExclude: [ExpandableList],
 			text: 'Basic usage of ExpandableList'
 		})(() => (
 			<ExpandableList
-				closeOnSelect={nullify(boolean('closeOnSelect', false))}
-				disabled={boolean('disabled', false)}
-				noAutoClose={nullify(boolean('noAutoClose', false))}
-				noLockBottom={nullify(boolean('noLockBottom', false))}
-				noneText={text('noneText', 'nothing selected')}
+				closeOnSelect={boolean('closeOnSelect', Config)}
+				disabled={boolean('disabled', Config)}
+				noAutoClose={boolean('noAutoClose', Config)}
+				noLockBottom={boolean('noLockBottom', Config)}
+				noneText={text('noneText', Config, 'nothing selected')}
 				onSelect={action('onSelect')}
 				onClose={action('onClose')}
 				onOpen={action('onOpen')}
-				select={select('select', ['radio', 'multiple', 'single'], 'radio')}
-				title={text('title', 'title')}
+				select={select('select', ['radio', 'multiple', 'single'], Config, 'radio')}
+				title={text('title', Config, 'title')}
 			>
 				{['option1', 'option2', 'option3']}
 			</ExpandableList>