+++ packages/sampler/stories/moonstone-stories/RangePicker.js
@@ -3,11 +3,11 @@ import {decrementIcons, incrementIcons} from './icons';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, number, select} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
+import {boolean, number, select} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 import nullify from '../../src/utils/nullify.js';
-import {mergeComponentMetadata} from '../../src/utils/propTables';
 
 const Config = mergeComponentMetadata('RangePicker', RangePickerBase, RangePicker);
 
@@ -24,27 +24,29 @@ const parseIntOrNullify = (v) => {
 	}
 };
 
+RangePicker.displayName = 'RangePicker';
+
 storiesOf('Moonstone', module)
 	.add(
 		'RangePicker',
 		withInfo({
-			propTables: [Config],
+			propTablesExclude: [RangePicker],
 			text: 'Basic usage of RangePicker'
 		})(() => (
 			<RangePicker
 				onChange={action('onChange')}
-				min={number('min', 0)}
-				max={number('max', 100)}
-				step={number('step', 5)}
+				min={number('min', Config, 0)}
+				max={number('max', Config, 100)}
+				step={number('step', Config, 5)}
 				defaultValue={0}
-				width={parseIntOrNullify(select('width', prop.width, 'small'))}
-				orientation={select('orientation', prop.orientation, 'horizontal')}
-				wrap={nullify(boolean('wrap', false))}
-				joined={nullify(boolean('joined', false))}
-				noAnimation={nullify(boolean('noAnimation', false))}
-				disabled={boolean('disabled', false)}
-				incrementIcon={nullify(select('incrementIcon', ['', ...incrementIcons]))}
-				decrementIcon={nullify(select('decrementIcon', ['', ...decrementIcons]))}
+				width={parseIntOrNullify(select('width', prop.width, Config, 'small'))}
+				orientation={select('orientation', prop.orientation, Config, 'horizontal')}
+				wrap={boolean('wrap', Config)}
+				joined={boolean('joined', Config)}
+				noAnimation={boolean('noAnimation', Config)}
+				disabled={boolean('disabled', Config)}
+				incrementIcon={select('incrementIcon', ['', ...incrementIcons], Config)}
+				decrementIcon={select('decrementIcon', ['', ...decrementIcons], Config)}
 			/>
 		))
 	);