+++ packages/sampler/stories/moonstone-stories/VirtualGridList.js
@@ -1,18 +1,15 @@
-import {VirtualGridList as UiVirtualGridList, VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList';
-import {VirtualGridList, VirtualListBase} from '@enact/moonstone/VirtualList';
+import {VirtualGridList as UiVirtualGridList} from '@enact/ui/VirtualList';
+import {VirtualGridList} from '@enact/moonstone/VirtualList';
 import {GridListImageItem as UiGridListImageItem} from '@enact/ui/GridListImageItem';
 import {GridListImageItem} from '@enact/moonstone/GridListImageItem';
 import ri from '@enact/ui/resolution';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {action} from '@storybook/addon-actions';
-import {boolean, number, select} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
-import nullify from '../../src/utils/nullify.js';
-import {mergeComponentMetadata} from '../../src/utils/propTables';
-
-const Config = mergeComponentMetadata('VirtualGridList', VirtualGridList, VirtualListBase, UiVirtualListBase);
+import {boolean, number, select} from '../../src/enact-knobs';
+import {mergeComponentMetadata} from '../../src/utils';
 
 const
 	wrapOption = {
@@ -75,24 +72,28 @@ const updateDataSize = (dataSize) => {
 
 updateDataSize(defaultDataSize);
 
+
+const Config = mergeComponentMetadata('VirtualGridList', UiVirtualGridList, VirtualGridList);
+UiVirtualGridList.displayName = 'VirtualGridList';
+
 storiesOf('UI', module)
 	.add(
 		'VirtualList.VirtualGridList',
 		withInfo({
-			propTables: [Config],
+			propTablesExclude: [UiVirtualGridList],
 			text: 'Basic usage of VirtualGridList'
 		})(() => (
 			<UiVirtualGridList
-				dataSize={updateDataSize(number('dataSize', defaultDataSize))}
-				direction={select('direction', prop.direction, 'vertical')}
+				dataSize={updateDataSize(number('dataSize', Config, defaultDataSize))}
+				direction={select('direction', prop.direction, Config, 'vertical')}
 				itemRenderer={uiRenderItem}
 				itemSize={{
-					minWidth: ri.scale(number('minWidth', 180)),
-					minHeight: ri.scale(number('minHeight', 270))
+					minWidth: ri.scale(number('minWidth', Config, 180)),
+					minHeight: ri.scale(number('minHeight', Config, 270))
 				}}
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
-				spacing={ri.scale(number('spacing', 20))}
+				spacing={ri.scale(number('spacing', Config, 20))}
 				style={{
 					height: ri.unit(549, 'rem')
 				}}
@@ -104,25 +105,25 @@ storiesOf('Moonstone', module)
 	.add(
 		'VirtualList.VirtualGridList',
 		withInfo({
-			propTables: [Config],
+			propTablesExclude: [VirtualGridList],
 			text: 'Basic usage of VirtualGridList'
 		})(() => (
 			<VirtualGridList
-				dataSize={updateDataSize(number('dataSize', defaultDataSize))}
-				direction={select('direction', prop.direction, 'vertical')}
-				focusableScrollbar={nullify(boolean('focusableScrollbar', false))}
+				dataSize={updateDataSize(number('dataSize', Config, defaultDataSize))}
+				direction={select('direction', prop.direction, Config, 'vertical')}
+				focusableScrollbar={boolean('focusableScrollbar', Config)}
 				itemRenderer={renderItem}
 				itemSize={{
-					minWidth: ri.scale(number('minWidth', 180)),
-					minHeight: ri.scale(number('minHeight', 270))
+					minWidth: ri.scale(number('minWidth', Config, 180)),
+					minHeight: ri.scale(number('minHeight', Config, 270))
 				}}
 				onScrollStart={action('onScrollStart')}
 				onScrollStop={action('onScrollStop')}
-				spacing={ri.scale(number('spacing', 20))}
+				spacing={ri.scale(number('spacing', Config, 20))}
 				style={{
 					height: ri.unit(549, 'rem')
 				}}
-				wrap={wrapOption[select('wrap', ['false', 'true', "'noAnimation'"])]}
+				wrap={wrapOption[select('wrap', ['false', 'true', "'noAnimation'"], Config)]}
 			/>
 		))
 	);